Converted most (probably) items to use users

+++ filmlog/filmstock.py
@@ -1,6 +1,8 @@
-from flask import request, render_template, redirect, url_for
+from flask import request, render_template, redirect, url_for, abort
 from sqlalchemy.sql import select, text, func
 import os, re
+
+from flask_login import LoginManager, login_required, current_user
 
 from filmlog import app
 from filmlog import database
@@ -9,38 +11,47 @@ from filmlog import functions
 engine = database.engine
 
 @app.route('/filmstock',  methods = ['GET', 'POST'])
+@login_required
 def filmstock():
     if request.method == 'POST':
         if request.form.get('button') == 'increment':
             if request.form.get('filmTypeID') != '' and request.form.get('filmSize') != '':
                 qry = text("""UPDATE FilmStock SET qty = qty + 1
                     WHERE filmTypeID = :filmTypeID
-                    AND filmSize = :filmSize""")
+                    AND filmSize = :filmSize
+                    AND userID = :userID""")
                 result = engine.execute(qry,
                     filmTypeID=request.form.get('filmTypeID'),
-                    filmSize=request.form.get('filmSize'))
+                    filmSize=request.form.get('filmSize'),
+                    userID = current_user.get_id())
         if request.form.get('button') == 'decrement':
             if request.form.get('filmTypeID') != '' and request.form.get('filmSize') != '':
                 qry = text("""SELECT qty FROM FilmStock
                     WHERE filmTypeID = :filmTypeID
-                    AND filmSize = :filmSize""")
+                    AND filmSize = :filmSize
+                    AND userID = :userID""")
                 result = engine.execute(qry,
                     filmTypeID=request.form.get('filmTypeID'),
-                    filmSize=request.form.get('filmSize')).fetchone()
+                    filmSize=request.form.get('filmSize'),
+                    userID = current_user.get_id()).fetchone()
                 if result.qty == 1:
                     qry = text(""" DELETE FROM FilmStock
                         WHERE filmTypeID = :filmTypeID
-                        AND filmSize = :filmSize""")
+                        AND filmSize = :filmSize
+                        AND userID = :userID""")
                     engine.execute(qry,
                         filmTypeID=request.form.get('filmTypeID'),
-                        filmSize=request.form.get('filmSize'))
+                        filmSize=request.form.get('filmSize'),
+                        userID = current_user.get_id())
                 else:
                     qry = text("""UPDATE FilmStock SET qty = qty - 1
                         WHERE filmTypeID = :filmTypeID
-                        AND filmSize = :filmSize""")
+                        AND filmSize = :filmSize
+                        AND userID = :userID""")
                     result = engine.execute(qry,
                         filmTypeID=request.form.get('filmTypeID'),
-                        filmSize=request.form.get('filmSize'))
+                        filmSize=request.form.get('filmSize'),
+                        userID = current_user.get_id())
         if request.form.get('button') == 'add':
             qty = request.form.get('qty')
             if request.form.get('filmTypeID') != '':
@@ -48,11 +59,12 @@ def filmstock():
                     qty = int(qty)
                     if qty > 0:
                         qry = text("""REPLACE INTO FilmStock
-                            (filmTypeID, filmSize, qty)
-                            VALUES (:filmTypeID, :filmSize, :qty)""")
+                            (filmTypeID, filmSize, userID, qty)
+                            VALUES (:filmTypeID, :filmSize, :userID, :qty)""")
                         result = engine.execute(qry,
                             filmTypeID=request.form.get('filmTypeID'),
                             filmSize=request.form.get('filmSize'),
+                            userID = current_user.get_id(),
                             qty=request.form.get('qty'))
     qry = text("""SELECT FilmStock.filmTypeID AS filmTypeID, filmSize, qty,
         FilmBrands.brand AS brand, FilmTypes.name AS type, iso
@@ -60,8 +72,9 @@ def filmstock():
         JOIN FilmTypes ON FilmTypes.filmTypeID = FilmStock.filmTypeID
         JOIN FilmBrands ON FilmBrands.filmBrandID = FilmTypes.filmBrandID
         WHERE filmSize IN ("35mm 24", "35mm 36", "35mm 100' Bulk Roll")
+        AND userID = :userID
         ORDER BY filmSize, brand, type, iso""")
-    stock_35mm = engine.execute(qry).fetchall()
+    stock_35mm = engine.execute(qry, userID = current_user.get_id()).fetchall()
 
     qry = text("""SELECT FilmStock.filmTypeID AS filmTypeID, filmSize, qty,
         FilmBrands.brand AS brand, FilmTypes.name AS type, iso
@@ -69,8 +82,9 @@ def filmstock():
         JOIN FilmTypes ON FilmTypes.filmTypeID = FilmStock.filmTypeID
         JOIN FilmBrands ON FilmBrands.filmBrandID = FilmTypes.filmBrandID
         WHERE filmSize IN ('120', '220')
+        AND userID = :userID
         ORDER BY filmSize, brand, type, iso""")
-    stock_mf = engine.execute(qry).fetchall()
+    stock_mf = engine.execute(qry, userID = current_user.get_id()).fetchall()
 
     qry = text("""SELECT FilmStock.filmTypeID AS filmTypeID, filmSize, qty,
         FilmBrands.brand AS brand, FilmTypes.name AS type, iso
@@ -78,8 +92,9 @@ def filmstock():
         JOIN FilmTypes ON FilmTypes.filmTypeID = FilmStock.filmTypeID
         JOIN FilmBrands ON FilmBrands.filmBrandID = FilmTypes.filmBrandID
         WHERE filmSize IN ('4x5', '8x10')
+        AND userID = :userID
         ORDER BY filmSize, brand, type, iso""")
-    stock_sheets = engine.execute(qry).fetchall()
+    stock_sheets = engine.execute(qry, userID = current_user.get_id()).fetchall()
 
     qry = text("""SELECT FilmTypes.filmTypeID AS filmTypeID,
         FilmBrands.brand AS brand, FilmTypes.name AS type, iso
